Move general purpose phake/functions.php
is_tty() moved to term_colors.php
resolve_runfile() and load_runfile() moved to phake/Builder.php
fatal() moved to phake/Bin.php
