+++ lib/includes.php
@@ -15,5 +15,4 @@ require_once __DIR__ . '/phake/OptionParser.php';
 require_once __DIR__ . '/phake/Task.php';
 require_once __DIR__ . '/phake/TaskCollisionException.php';
 require_once __DIR__ . '/phake/TaskNotFoundException.php';
-require_once __DIR__ . '/phake/functions.php';
 require_once __DIR__ . '/phake/Utils.php';
